requirements.txt
fpdf==1.7.2
httplib2==0.22.0
PyAfipWs==3.8.0
pyparsing==3.1.1
PySimpleSOAP==1.8.14
python-dotenv==1.0.0
